+++ packages/components/addon/components/hds/dropdown/list-item/checkbox.js
@@ -4,7 +4,7 @@
  */
 
 import Component from '@glimmer/component';
-import { getElementId } from '@hashicorp/design-system-components/helpers/hds-get-element-id';
+import { getElementId } from '@hashicorp/design-system-components/utils/hds-get-element-id';
 
 export default class HdsDropdownListItemCheckboxComponent extends Component {
   /**
