+++ packages/components/addon/components/hds/form/field/index.js
@@ -8,8 +8,8 @@ import { tracked } from '@glimmer/tracking';
 import { assert } from '@ember/debug';
 import { action } from '@ember/object';
 import { schedule } from '@ember/runloop';
-import { getElementId } from '@hashicorp/design-system-components/helpers/hds-get-element-id';
-import { setAriaDescribedBy } from '@hashicorp/design-system-components/helpers/hds-set-aria-described-by';
+import { getElementId } from '@hashicorp/design-system-components/utils/hds-get-element-id';
+import { setAriaDescribedBy } from '@hashicorp/design-system-components/utils/hds-set-aria-described-by';
 
 export const LAYOUT_TYPES = ['vertical', 'flag'];
