+++ packages/components/app/utils/hds-set-aria-described-by.js
@@ -1,0 +1,6 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+export { default } from '@hashicorp/design-system-components/utils/hds-set-aria-described-by';
